+++ lib/DAV/Sync/Plugin.php
@@ -333,5 +333,26 @@ class Plugin extends DAV\ServerPlugin {
 
     }
 
+    /**
+     * Returns a bunch of meta-data about the plugin.
+     *
+     * Providing this information is optional, and is mainly displayed by the
+     * Browser plugin.
+     *
+     * The description key in the returned array may contain html and will not
+     * be sanitized.
+     *
+     * @return array
+     */
+    function getPluginInfo() {
+
+        return [
+            'name'        => $this->getPluginName(),
+            'description' => 'Adds support for WebDAV Collection Sync (rfc6578)',
+            'link'        => 'http://sabre.io/dav/sync/',
+        ];
+
+    }
+
 }